pyproject.toml
[project]
name = "aqua-whisper"
version = "0.1.0"
description = "FastAPI + Celery + Redis async YouTube transcript pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "celery[redis]",
    "redis",
    "httpx",
    "pydantic-settings",
    "faster-whisper",
    "yt-dlp",
    "structlog",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
